backend/requirements.txt
albumentations==1.4.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
dlib @ file:///Users/runner/miniforge3/conda-bld/dlib_1697537211922/work
easydict==1.13
exceptiongroup==1.2.2
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.5.3
fastapi==0.95.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gdown==4.7.1
h11==0.14.0
httptools==0.6.4
idna==3.10
imageio==2.36.0
importlib_resources==6.4.5
insightface==0.7.3
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.4
onnx==1.17.0
onnxruntime-silicon==1.11.1
opencv-python==4.10.0.84
opencv-python-headless==4.8.0.74
packaging==24.2
pandas==1.5.3
pillow==11.0.0
prettytable==3.12.0
protobuf==5.28.3
pydantic==1.10.19
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1731919281354/work
python-dotenv==1.0.1
python-multipart==0.0.6
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1726055524169/work
PyYAML==6.0.2
qudida==0.0.4
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
soupsieve==2.6
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.5.1
torchvision==0.20.1
tqdm==4.65.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
zipp==3.21.0
